Completes migration of files to csv module, modified region module to use it
