+++ csv/src/main/java/org/motechproject/nms/csv/utils/ConstraintViolationUtils.java
@@ -1,4 +1,4 @@
-package org.motechproject.nms.region.utils;
+package org.motechproject.nms.csv.utils;
 
 import javax.validation.ConstraintViolation;
 import java.util.Set;
