+++ csv/src/main/java/org/motechproject/nms/csv/utils/CsvImporter.java
@@ -1,4 +1,4 @@
-package org.motechproject.nms.region.utils;
+package org.motechproject.nms.csv.utils;
 
 import org.supercsv.cellprocessor.ift.CellProcessor;
 import org.supercsv.io.ICsvReader;
