+++ csv/src/main/java/org/motechproject/nms/csv/utils/CsvMapImporter.java
@@ -1,4 +1,4 @@
-package org.motechproject.nms.region.utils;
+package org.motechproject.nms.csv.utils;
 
 import org.supercsv.io.CsvMapReader;
 import org.supercsv.io.ICsvMapReader;
