+++ csv/src/main/java/org/motechproject/nms/csv/utils/GetBoolean.java
@@ -1,7 +1,7 @@
-package org.motechproject.nms.region.utils;
+package org.motechproject.nms.csv.utils;
 
 import org.apache.commons.lang.StringUtils;
-import org.motechproject.nms.region.exception.CsvImportDataException;
+import org.motechproject.nms.csv.exception.CsvImportDataException;
 import org.supercsv.cellprocessor.ift.CellProcessor;
 import org.supercsv.util.CsvContext;
