+++ csv/src/main/java/org/motechproject/nms/csv/utils/GetInstanceByLong.java
@@ -1,4 +1,4 @@
-package org.motechproject.nms.region.utils;
+package org.motechproject.nms.csv.utils;
 
 import org.supercsv.util.CsvContext;
