+++ csv/src/main/java/org/motechproject/nms/csv/utils/GetInteger.java
@@ -1,19 +1,19 @@
-package org.motechproject.nms.region.utils;
+package org.motechproject.nms.csv.utils;
 
-import org.motechproject.nms.region.exception.CsvImportDataException;
+import org.motechproject.nms.csv.exception.CsvImportDataException;
 import org.supercsv.cellprocessor.ift.CellProcessor;
 import org.supercsv.util.CsvContext;
 
-public class GetLong implements CellProcessor {
+public class GetInteger implements CellProcessor {
 
     @Override
     public Object execute(Object value, CsvContext context) {
-        Long returnValue;
+        Integer returnValue;
         if (value instanceof Number) {
-            returnValue = ((Number) value).longValue();
+            returnValue = ((Number) value).intValue();
         } else if (value instanceof String) {
             try {
-                returnValue = Long.valueOf((String) value);
+                returnValue = Integer.valueOf((String) value);
             } catch (NumberFormatException e) {
                 throw new CsvImportDataException(getErrorMessage(value, context), e);
             }
@@ -24,7 +24,7 @@ public class GetLong implements CellProcessor {
     }
 
     private String getErrorMessage(Object value, CsvContext context) {
-        return String.format("CSV field error [row: %d, col: %d]: Expected Long value, found %s",
+        return String.format("CSV field error [row: %d, col: %d]: Expected Integer value, found %s",
                 context.getRowNumber(), context.getColumnNumber(), value);
     }
