+++ csv/src/main/java/org/motechproject/nms/csv/utils/GetString.java
@@ -1,6 +1,6 @@
-package org.motechproject.nms.region.utils;
+package org.motechproject.nms.csv.utils;
 
-import org.motechproject.nms.region.exception.CsvImportDataException;
+import org.motechproject.nms.csv.exception.CsvImportDataException;
 import org.supercsv.cellprocessor.ift.CellProcessor;
 import org.supercsv.util.CsvContext;
