+++ csv/src/test/java/org/motechproject/nms/csv/ut/CsvImportTest.java
@@ -4,14 +4,14 @@ import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
 import org.motechproject.mds.service.MotechDataService;
-import org.motechproject.nms.region.exception.CsvImportDataException;
-import org.motechproject.nms.region.utils.CsvInstanceImporter;
-import org.motechproject.nms.region.utils.CsvMapImporter;
-import org.motechproject.nms.region.utils.GetBoolean;
-import org.motechproject.nms.region.utils.GetInstanceByLong;
-import org.motechproject.nms.region.utils.GetInteger;
-import org.motechproject.nms.region.utils.GetLong;
-import org.motechproject.nms.region.utils.GetString;
+import org.motechproject.nms.csv.exception.CsvImportDataException;
+import org.motechproject.nms.csv.utils.CsvInstanceImporter;
+import org.motechproject.nms.csv.utils.CsvMapImporter;
+import org.motechproject.nms.csv.utils.GetBoolean;
+import org.motechproject.nms.csv.utils.GetInstanceByLong;
+import org.motechproject.nms.csv.utils.GetInteger;
+import org.motechproject.nms.csv.utils.GetLong;
+import org.motechproject.nms.csv.utils.GetString;
 import org.supercsv.cellprocessor.ift.CellProcessor;
 import org.supercsv.util.CsvContext;
