+++ pom.xml
@@ -727,6 +727,7 @@
     </dependencyManagement>
 
     <modules>
+        <module>csv</module>
         <module>mobile-academy</module>
         <module>kilkari</module>
         <module>flw</module>
