+++ region/pom.xml
@@ -48,6 +48,12 @@
         </dependency>
 
         <dependency>
+            <groupId>org.motechproject.nms</groupId>
+            <artifactId>csv</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
+        <dependency>
             <groupId>commons-fileupload</groupId>
             <artifactId>commons-fileupload</artifactId>
         </dependency>
@@ -87,9 +93,10 @@
                             org.motechproject.nms.region.service;version=${project.version},
                             org.motechproject.nms.region.repository;version=${project.version},
                             org.motechproject.nms.region.domain;version=${project.version},
-                            org.motechproject.nms.region.exception;version=${project.version}
                         </Export-Package>
                         <Import-Package>
+                            org.motechproject.nms.csv.exception,
+                            org.motechproject.nms.csv.utils,
                             net.sf.cglib.core,
                             net.sf.cglib.proxy,
                             net.sf.cglib.reflect,
