--- region/src/main/java/org/motechproject/nms/region/exception/CsvImportException.java
@@ -1,14 +1,0 @@
-package org.motechproject.nms.region.exception;
-
-public class CsvImportException extends RuntimeException {
-
-    private static final long serialVersionUID = -6278097435395933061L;
-
-    public CsvImportException(String message) {
-        super(message);
-    }
-
-    public CsvImportException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
