+++ region/src/main/java/org/motechproject/nms/region/service/impl/HealthFacilityImportServiceImpl.java
@@ -7,9 +7,9 @@ import org.motechproject.nms.region.repository.HealthBlockDataService;
 import org.motechproject.nms.region.repository.HealthFacilityDataService;
 import org.motechproject.nms.region.repository.HealthFacilityTypeDataService;
 import org.motechproject.nms.region.service.HealthFacilityImportService;
-import org.motechproject.nms.region.utils.GetInstanceByLong;
-import org.motechproject.nms.region.utils.GetLong;
-import org.motechproject.nms.region.utils.GetString;
+import org.motechproject.nms.csv.utils.GetInstanceByLong;
+import org.motechproject.nms.csv.utils.GetLong;
+import org.motechproject.nms.csv.utils.GetString;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.supercsv.cellprocessor.ift.CellProcessor;
