+++ region/src/main/java/org/motechproject/nms/region/service/impl/HealthSubFacilityImportServiceImpl.java
@@ -5,9 +5,9 @@ import org.motechproject.nms.region.domain.HealthSubFacility;
 import org.motechproject.nms.region.repository.HealthFacilityDataService;
 import org.motechproject.nms.region.repository.HealthSubFacilityDataService;
 import org.motechproject.nms.region.service.HealthSubFacilityImportService;
-import org.motechproject.nms.region.utils.GetInstanceByLong;
-import org.motechproject.nms.region.utils.GetLong;
-import org.motechproject.nms.region.utils.GetString;
+import org.motechproject.nms.csv.utils.GetInstanceByLong;
+import org.motechproject.nms.csv.utils.GetLong;
+import org.motechproject.nms.csv.utils.GetString;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.supercsv.cellprocessor.ift.CellProcessor;
