+++ region/src/main/java/org/motechproject/nms/region/service/impl/NonCensusVillageImportServiceImpl.java
@@ -5,9 +5,9 @@ import org.motechproject.nms.region.domain.Village;
 import org.motechproject.nms.region.repository.TalukaDataService;
 import org.motechproject.nms.region.repository.VillageDataService;
 import org.motechproject.nms.region.service.NonCensusVillageImportService;
-import org.motechproject.nms.region.utils.GetInstanceByString;
-import org.motechproject.nms.region.utils.GetLong;
-import org.motechproject.nms.region.utils.GetString;
+import org.motechproject.nms.csv.utils.GetInstanceByString;
+import org.motechproject.nms.csv.utils.GetLong;
+import org.motechproject.nms.csv.utils.GetString;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.supercsv.cellprocessor.Optional;
